requirement.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-oauth==1.1.0
sqlparse==0.5.4
tzdata==2025.3
